Everything that is not commented out works. So please run these smoke tests after introducing changes to SDM. I'll work on something that does that automatically next. Topics are indeed created now, so that has been uncommented. Commented out tests or steps are either undefined or result in an error. Specifically there are still issues with local deploy and approving goals. The immaterial behavior went away, so those tests have been commented out too. I could add back behavior that would make readme.md changes immaterial. From what I understand, there was a decision to go from whitelist to blacklist.
The node repo used in tests is now buildable-node-seed so that a build really builds and can succeed or fail.

I also got sidetracked with reviews not working. They seem to take longer than they used to, and they were failing. The log was not help, but debugging into the node_modules sdm compiled javascript, I could. see that it was failing on the cloud readiness review listener. It could not create the issue stating that losgatos1 was not cloud ready. This was because losgatos1 was forked into my org, and apparently issues are not enabled. This results in a 410 http code. I found the call, and made the same one manually, and then I could see the message. It was a quick change to enable issues for the repo, but debuggin this way is time consuming.

Currently there are some secret requirements on the org you want to run smoke tests against. I have created https://github.com/atomist-fidelity just for running smoke tests. In some tests the repo is created, but some count on a repo already existing. So you must fork spring-team/java-hello-world-maven and spring-team/losgatos1. The latter must have issues enabled.