fixed tests to reflect reality around local deployment
* deploy locally no longer checks the endpoint, so that assertion was removed
This could have been an error because it is confusing to have LocalDeploymentGoal and two LocalDeploymentGoals. I recommend we use the LocalDeploymentGoals in the sdm which is identical to the one in sample-sdm, and maybe name the grouped goal differently. I'm assuming that no endpoint assertion is the new intended behavior.

* deploy locally was removed from springBootCreation.feature because it now only happens on branches.